feat: events parsing from message logs wip
